// src/Types/TypeVechiles.tsx
export type VechilesSearchType = {
    cargo_capacity: string
    consumables: string
    cost_in_credits: string
    created: string
    crew: string
    edited: string
    length: string
    manufacturer: string
    max_atmosphering_speed: string
    model: string
    name: string
    passengers: string
    pilots: string
    films: string
    url: string
    vehicle_class: string
}

export type VechilesSpecificType = {
    cargo_capacity: string
    consumables: string
    cost_in_credits: string
    created: string
    crew: string
    edited: string
    length: string
    manufacturer: string
    max_atmosphering_speed: string
    model: string
    name: string
    passengers: string
    pilots: string
    films: string
    url: string
    vehicle_class: string
}